Merge Frontend

+++ backend/app/crud/manager_crud.py
@@ -1,14 +1,13 @@
 from typing import Optional, List
 from sqlalchemy.orm import Session
 from sqlalchemy import select
-
+from app.models.role_model import Role 
 from app.models.manager_model import Manager
 from app.schemas.manager_schema import ManagerUpdate, ManagerCreate
-from app.schemas.user_role_schema import ManagerCreateWithUser
 from app.models.association_tables import user_roles 
 
-def get_manager(db: Session, manager_id: int) -> Optional[Manager]:
-    stmt = select(Manager).where(Manager.manager_id == manager_id)
+def get_manager(db: Session, user_id: int) -> Optional[Manager]:
+    stmt = select(Manager).where(Manager.user_id == user_id)
     return db.execute(stmt).scalar_one_or_none()
 
 def get_manager_by_user_id(db: Session, user_id: int) -> Optional[Manager]:
@@ -29,8 +28,8 @@ def create_manager(db: Session, manager_in: ManagerCreate) -> Manager:
     db.refresh(db_manager)
     return db_manager
 
-def update_manager(db: Session, manager_id: int, manager_update: ManagerUpdate) -> Optional[Manager]:
-    db_manager = get_manager(db, manager_id)
+def update_manager(db: Session, user_id: int, manager_update: ManagerUpdate) -> Optional[Manager]:
+    db_manager = get_manager(db, user_id)
     if not db_manager:
         return None
     update_data = manager_update.model_dump(exclude_unset=True)
@@ -40,16 +39,25 @@ def update_manager(db: Session, manager_id: int, manager_update: ManagerUpdate) 
     db.refresh(db_manager)
     return db_manager
 
-def delete_manager(db: Session, manager_id: int):
-    db_manager = db.query(Manager).filter(Manager.manager_id == manager_id).first()
+def delete_manager(db: Session, user_id: int):
+    db_manager = get_manager(db, user_id=user_id)
     if not db_manager:
         return None
-    # Lưu dữ liệu trước khi xóa
-    deleted_data = db_manager
+
+    # Lấy role "manager"
+    manager_role = db.query(Role).filter(Role.name == "manager").first()
+    if not manager_role:
+        return None
+
+    # Xóa chỉ role "manager" cho user này
     db.execute(
-        user_roles.delete().where(user_roles.c.user_id == db_manager.user_id)
+        user_roles.delete()
+        .where(user_roles.c.user_id == db_manager.user_id)
+        .where(user_roles.c.role_id == manager_role.role_id)
     )
+
     db.delete(db_manager)
     db.commit()
-    return deleted_data
+    return db_manager
+
